Streamlit_chatapp-main/requirements.txt
langchain
streamlit
python-dotenv
openai
st-chat-message
pandas